+++ src/data/steps.js
@@ -2,7 +2,7 @@ import {
   Heart,
   FolderUp,
   UserRound,
-  ClipboardList,
+  DollarSign,
   GraduationCap,
   ClipboardCheck,
   MonitorPlay,
@@ -24,12 +24,13 @@ const nurseSteps = [
   { id: 2, stepName: "Upload Documents", logo: FolderUp },
   { id: 3, stepName: "Personal Information", logo: UserRound },
   { id: 4, stepName: "New Hire Orientation", logo: GraduationCap },
-  { id: 5, stepName: "Bloodborne Pathogens", logo: HeartPulse},
-  { id: 6, stepName: "How to Use eRSP", logo: MonitorPlay },
+  { id: 5, stepName: "How to Use eRSP", logo: MonitorPlay },
+  { id: 6, stepName: "How to Use SurePayroll", logo: DollarSign },
   { id: 7, stepName: "Forms & Agreements", logo: ScrollText },
   { id: 8, stepName: "Tax Forms (W-9)", logo: FileText },
   { id: 9, stepName: "Offer Letter", logo: FileSignature },
-  { id: 10, stepName: "Completed!", logo: PartyPopper },
+  { id: 10, stepName: "Bloodborne Pathogens", logo: HeartPulse },
+  { id: 11, stepName: "Completed!", logo: PartyPopper },
 ]
 
 const nurseWelcomeSteps = [
@@ -49,13 +50,14 @@ export const stepsByRole = {
     { id: 2, stepName: "Upload Documents", logo: FolderUp },
     { id: 3, stepName: "Personal Information", logo: UserRound },
     { id: 4, stepName: "New Hire Orientation", logo: GraduationCap },
-    { id: 5, stepName: "Bloodborne Pathogens", logo: HeartPulse },
-    { id: 6, stepName: "Competency Checklist", logo: ClipboardCheck },
-    { id: 7, stepName: "How to Use eRSP", logo: MonitorPlay },
+    { id: 5, stepName: "Competency Checklist", logo: ClipboardCheck },
+    { id: 6, stepName: "How to Use eRSP", logo: MonitorPlay },
+    { id: 7, stepName: "How to Use SurePayroll", logo: DollarSign },
     { id: 8, stepName: "Forms & Agreements", logo: ScrollText },
     { id: 9, stepName: "Tax Forms", logo: FileText },
     { id: 10, stepName: "Offer Letter", logo: FileSignature },
-    { id: 11, stepName: "Completed!", logo: PartyPopper },
+    { id: 11, stepName: "Bloodborne Pathogens", logo: HeartPulse },
+    { id: 12, stepName: "Completed!", logo: PartyPopper },
   ]),
 
   nurse_prn: makeSteps(nurseSteps),
@@ -66,11 +68,12 @@ export const stepsByRole = {
     { id: 2, stepName: "Upload Documents", logo: FolderUp },
     { id: 3, stepName: "Personal Information", logo: UserRound },
     { id: 4, stepName: "New Hire Orientation", logo: GraduationCap },
-    { id: 5, stepName: "Bloodborne Pathogens", logo: HeartPulse},
+    { id: 5, stepName: "How to Use SurePayroll", logo: DollarSign },
     { id: 6, stepName: "Forms & Agreements", logo: ScrollText },
     { id: 7, stepName: "Tax Forms", logo: FileText },
     { id: 8, stepName: "Offer Letter", logo: FileSignature },
-    { id: 9, stepName: "Completed!", logo: PartyPopper },
+    { id: 9, stepName: "Bloodborne Pathogens", logo: HeartPulse },
+    { id: 10, stepName: "Completed!", logo: PartyPopper },
   ]),
 };
 
@@ -80,7 +83,7 @@ export const welcomeSteps = {
     "✓ Personal information & emergency contact information",
     "✓ New hire orientation",
     "✓ Competency checklist",
-    "✓ Guide to using eRSP",
+    "✓ Guide to using eRSP & SurePayroll",
     "✓ Forms & agreements",
     "✓ Tax forms",
     "✓ Offer letter"
@@ -92,6 +95,7 @@ export const welcomeSteps = {
     "✓ Personal information & emergency contact",
     "✓ New hire orientation",
     "✓ Forms & agreements",
+    "✓ Guide to using SurePayroll",
     "✓ Tax forms",
     "✓ Offer letter"
   ]